+++ frontend/src/hooks/index.ts
@@ -1,4 +1,4 @@
-import useIntersectionObserver from 'src/hooks/IntersectionObserver';
+import useIntersectionObserver from 'src/hooks/intersectionObserver';
 
 // eslint-disable-next-line import/prefer-default-export
 export { useIntersectionObserver };
