+++ frontend/src/hooks/intersectionObserver.ts
